PostgreSQL/requirements.txt
google-api-python-client==2.15.0
oauth2client==4.1.3
apiclient==1.0.4
httplib2==0.19.1
psycopg2-binary==2.9.1
